requirements.txt
fastapi
uvicorn
sentry_sdk
python-dotenv
python-multipart
langchain
langchain-community==0.0.37
databases[postgresql]
pydantic
python-dotenv
google-cloud-storage
google-auth
boto3
botocore
azure-storage-blob
sqlalchemy
psycopg2-binary
qdrant-client
openai
alembic
huggingface_hub
pgvector
psycopg[binary,pool]
